// samples/file-browsing/typescript-react/src/types.ts
export interface IAuthenticateCommand {
    command: "authenticate";
    resource: string;
    type: "SharePoint" | "Graph";
    claimsChallenge?: {
        claims: string;
        token?: string;
        error: any; // IError;
    };
    expiration?: {
        error: any; // IError;
    };
}

export type IThemeName =
    | "default"
    | "dark"
    | "lists"
    | "teams"
    | "teams-default"
    | "teams-dark"
    | "teams-contrast";

export interface ILayerConfiguration {
    mode?: "z-index" | "projection-frame" | "default";
    offset?: {
        top?: number;
        left?: number;
    };
}

export interface IBreadcrumbConfiguration {
    /**
     * Indicate whether set the root crumb folder path to the entry folder specified in IEntryConfiguration
     * TODO: deprecated. Should use root to specific the actual root crumb folder path.
     */
    setRootToEntry?: boolean;

    /**
     * Specify the root crumb folder path
     */
    root?: {
        sharePoint?: {
            byPath?: {
                web?: string;
                list?: string;
                folder?: string;
            };
        };
    };
}

/**
 * Configuration for communication between the host and the experience.
 * This must be specified in order to authenticate the experience, handle button clicks,
 * and receive data back.
 */
export interface IMessagingConfiguration {
    /**
     * A unique id assigned by the host app to this File Browser instance.
     * This should ideally be a new GUID generated by the host.
     */
    channelId: string;
    /**
     * The host app"s authority, used as the target origin for post-messaging.
     */
    origin: string;
    /**
     * Whether or not the host app window will need to identify itself.
     */
    identifyParent?: boolean;
    /**
     * Whether or not the client app must wait for a "configure" command to be sent by the host before rendering.
     */
    waitForConfiguration?: boolean;
    /**
     * Whether or not the client app can re-establish a connection to the host after a reload.
     */
    restartable?: boolean;
    /**
     * Override timeout for acknowledgement messages.
     */
    acknowledgeTimeout?: number;
    /**
     * Override timeout for the initialization handshake.
     */
    initializeTimeout?: number;
    /**
     * Override timeout for command responses.
     */
    resultTimeout?: number;
}

/**
 * Configures a pivot in the left-nav to behavior like a custom command.
 */
export interface IPivotCommandConfiguration {
    key?: string;
    action: "custom" | "open";
}

/**
 * Configures what types of items are allowed to be browsered, selected, and picked within the experience.
 * Note that the default configuration accounts for the expected authentication capabilities of the host app.
 * Depending on what else is enabled by the host, the host may be expected to provide tokens for more
 * services and scopes.
 */
export interface ITypesAndSourcesConfiguration {
    /**
     * Configures the base type of item which will be picked and displayed.
     * Only items of this general type will be selectable.
     */
    mode?: "files" | "folders" | "all";
    /**
     * `filters` options: file extension, i.e. .xlsx, .docx, .ppt, etc.
     * `filters` options: "photo", "folder", "video", "documentLibrary"
     */
    filters?: string[];
    /**
     * Specifies a filter for *where* the item may come from.
     */
    locations?: {
        /**
         * Items may only come from the user"s OneDrive.
         */
        oneDrive?: {};
        /**
         * Items may only come from a specific location within SharePoint.
         */
        sharePoint?: {
            byPath?: {
                web?: string;
                list?: string;
                folder?: string;
            };
        };
    };
    /**
     * Specifies which pivots the user may access while browsing files and lists.
     * Note that if a pivot is disabled here but still targeted in `entry`, it will still be visible
     * in the nav.
     */
    pivots?: {
        /**
         * Show "My files".
         */
        oneDrive?: boolean | IPivotCommandConfiguration;
        /**
         * Show "Photos". Consumer accounts only.
         */
        photos?: boolean | IPivotCommandConfiguration;
        /**
         * Show "Recent".
         */
        recent?: boolean | IPivotCommandConfiguration;
        /**
         * Show "Shared"
         */
        shared?: boolean | IPivotCommandConfiguration;
        /**
         * Show "Mobile device". The host app must be able to authenticate this service separately.
         */
        phoneLink?: boolean;
        /**
         * Show "Communities". The host app must be Yammer.
         */
        yammerCommunities?: boolean;
        /**
         * Show "Cloud storage". The host app must be Teams.
         */
        teamsCloudStorage?: boolean;
        /**
         * Show "Quick access".
         */
        sharedLibraries?: boolean;
        /**
         * Show "Stock images".
         */
        stockImages?: boolean;
        /**
         * Show "Bing search".
         */
        bing?: boolean;
        /**
         * Show "My lists".
         */
        myLists?: boolean;
    };
}

export interface IAccessibilityConfiguration {
    /**
     * Whether or not to "trap focus" within the component. If this is enabled, tab-stops will loop from
     * the last element back to the left navigation automatically. This is useful if the components"s frame
     * is hosted as the only content of a modal overlay and focus should not jump to the outside content.
     *
     * @default `false`
     */
    enableFocusTrap?: boolean;
    /**
     * Whether or not the component should immediately grab focus once the content has loaded.
     *
     * @default `true`
     */
    trapFocusOnLoad?: boolean;
    /**
     * Whether or not to force the currently-focused element within the component to be highlighted.
     * By default, the focused element is highlighted if the user navigates elements with the keyboard
     * but not when the user interacts via the mouse. However, if a host application launches the component
     * due to keyboard input is should set this flag to `true` to ensure continuity of behavior.
     *
     * @default `false`
     */
    showFocusOnLoad?: boolean;
}

export type FilePickerPickAction = "select" | "share" | "download" | "move";
// ...
export interface ICommandsConfiguration {
    /**
     * Overrides the "open" behavior for files and folders.
     */
    open?: {
        /**
         * Specifies a series of "handlers" for clicking files and folders.
         * For each handler, the invoked item is tested against `filters`,
         * and the first matching handler has its behavior applied.
         * If no handler matches, the default behavior applies.
         */
        handlers?: {
            filters?: string[];
            /**
             * Specifies the target for opening the item
             * - `none`: Do not allow the item to be opened.
             * - `navigate`: Open the item within the experience.
             * - `external`: Open the item in a new tab.
             * - `host`: Ask the host to open the item.
             */
            target?: "none" | "navigate" | "external" | "host";
        }[];
    };
    /**
     * Overrides behaviors for the view-switch dropdown.
     */
    switchView?: {
        enabled?: boolean;
        views: {
            smartTemplate?: boolean; // default false
        };
    };
    /**
     * Overrides behaviors for editing SharePoint views.
     */
    editView?: {
        enabled?: boolean;
    };
    /**
     * Overrides behaviors for creating SharePoint views.
     */
    createView?: {
        enabled?: boolean;
    };
    /**
     * Specifies the behavior for file-picking.
     */
    pick?: {
        /**
         * A special action to perform when picking the file, before handing the result
         * back to the host app.
         */
        action?: FilePickerPickAction;
        /**
         * A custom label to apply to the button which picks files.
         * This must be localized by the host app if supplied.
         */
        label?: string;
        /**
         * Configures the "move" action for picking files.
         */
        move?: {
            sourceItems?: IItem[];
        };
        /**
         * Configures the "copy" action for picking files.
         */
        copy?: {
            sourceItems?: IItem[];
        };
        /**
         * Configures the "select" action for picking files.
         */
        select?: IFilePickerPickCommandSelectOptions;
        /**
         * Configures special handling for custom lists.
         */
        listItem?: {
            fieldName?: string;
        };
    };
    /**
     * Specifies the behavior for closing the experience.
     */
    close?: {
        /**
         * A custom label to apply to the "cancel" button.
         * This must be localized by the host app if supplied.
         */
        label?: string;
    };
    /**
     * Behavior for a "Browse this device" command to pick local files.
     */
    browseThisDevice?: {
        enabled?: boolean;
        label?: string;
    };
    /**
     * Behavior for a "From a link" command to pick from a link.
     */
    fromALink?: {
        enabled?: boolean;
    };
    /**
     * Behavior for a "Switch account" command.
     */
    switchAccount?: {
        mode?: "host" | "none";
    };
    /**
     * Behavior for a "Manage accounts" command.
     */
    manageAccounts?: {
        mode?: "host" | "none";
        label?: string;
    };
    /**
     * Behavior for "Upload"
     */
    upload?: {
        enabled?: boolean;
    };
    /**
     * Behavior for "Create folder"
     */
    createFolder?: {
        enabled?: boolean;
    };
    /**
     * Behavior for "Filter by" in the column headers.
     */
    filterByColumn?: {
        mode?: "panel" | "menu";
    };
    /**
     * Configuration for a command to open the current item in the full SharePoint site.
     */
    openInSharePoint?: {
        enabled?: boolean;
        location?: "header" | "command-bar";
    };
    /**
     * How to handle "Open in Desktop" protocol callbacks
     */
    openInDesktop?: {
        /**
         * @default "inline"
         */
        mode: "external" | "inline" | "host" | "none";
    };
    /**
     * How to handle actions defined by custom formatters.
     */
    customFormatter?: {
        actions?: {
            key: string;
            mode?: "host" | "none";
        }[];
    };
    /**
     * How to handle specified values for `key` in custom commands
     * in the tray, nav, or command bar.
     */
    custom?: {
        actions?: ICustomCommandActionConfiguration[];
    };
}

export interface IFilePickerPickCommandSelectOptions {
    urls?: {
        download?: boolean;
    };
}
// ...
export type ListLayoutType = "details" | "compact-details" | "tiles";

export interface IListConfiguration {
    /**
     * A custom override for the initial list layout.
     */
    layout?: {
        /**
         * Sets the preferred starting layout for the initial content.
         */
        type?: ListLayoutType;
    };
    /**
     * Configures scrolling behavior within the Browser.
     */
    scrolling?: {
        enableStickyHeaders?: boolean;
    };
}

/**
 * Configuration for handling authentication requests from the embedded app.
 * Presence of this object (even if empty) indicates that the host will handle authentication.
  */
export interface IAuthenticationConfiguration {}
// ...
export interface ILeftNavConfiguration {
    /**
     * Whether or not a Left Nav should be rendered by the embedded content.
     */
    enabled?: boolean;
    /**
     * Mode of presentation of the nav.
     * If the nav is enabled but this is set to `host`, the embedded app
     * will show a button to ask the host app to show a nav.
     */
    mode?: "host" | "default";
    /**
     * Indicates whether the left nav will be initially modal.
     */
    initialModality?: "modal" | "hidden";

    /**
     * Type of left nav
     */
    preset?: "oneDrive" | "current-site";

    /**
     * Custom commands to insert at the end of the left nav. Will appear before the default set.
     */
    commands?: ILeftNavCommandConfiguration[];
}
// ...
export interface IFileBrowserOptions {
    sdk: "1.0" | "v1.0" | "8.0" | "v8.0";
    messaging: IMessagingConfiguration;
    entry?: IEntryConfiguration;
    authentication?: IAuthenticationConfiguration;
    breadcrumb?: IBreadcrumbConfiguration;
    commands?: ICommandsConfiguration;
    typesAndSources?: ITypesAndSourcesConfiguration;
    accessibility?: IAccessibilityConfiguration;
    leftNav?: ILeftNavConfiguration;
    list?: IListConfiguration;
    layer?: ILayerConfiguration;
}
// ...
export interface IItem {

}